% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotWrappers.R
\name{ggplotLatentMap}
\alias{ggplotLatentMap}
\title{ggplotLatentMap: function for plotting data and smoothed model next to each other on map}
\usage{
ggplotLatentMap(model, shp, title = "",
  shape_level = "residence_census_tract")
}
\arguments{
\item{model}{model object from incidenceMapR::modelTrainR}

\item{shp}{sf object with residence_census_tract shapes (residence_census_tract only for now)}
}
\value{
ggplot object
}
\description{
ggplotLatentMap: function for plotting data and smoothed model next to each other on map
}
\examples{
plotDat <- right_join(model$modeledData,shp, by=c('residence_census_tract'))
plotDat$positive[plotDat$positive==0]<-NaN
ggplotSmoothMap(plotDat,shp)

}
