% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotWrappers.R
\name{ggplotSmoothEffects}
\alias{ggplotSmoothEffects}
\title{ggplotSmoothEffects: function for plotting smoothed model and fixed effects model next to each other}
\usage{
ggplotSmoothEffects(model_smooth, model_effects, shp, censusdb = NULL,
  title = "", shape_level = "residence_census_tract")
}
\arguments{
\item{plotDat}{data.frame that joins dbViewR::modeledData}
}
\value{
ggplot object
}
\description{
ggplotSmoothEffects: function for plotting smoothed model and fixed effects model next to each other
}
\examples{


}
